+++ src-tauri/build.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 refinex-lab
+
+fn main() {
+    tauri_build::build()
+}